feat(rfc-0005): Fase 1 — migra módulos puros para src/hronir/ (TypeScript)
posts, matches, moods, perspectives, ranking migrados para src/hronir/ com
tipos explícitos; scripts/hronir/lib/*.js viram re-exports finos; teste
atualizado para mockar e importar de src/hronir/ direto.
