+++ package.json
@@ -34,7 +34,7 @@
     "hronir:migrate": "node --import tsx/esm scripts/hronir/index.js migrate",
     "hronir:doctor": "node --import tsx/esm scripts/hronir/index.js doctor",
     "music:generate": "node scripts/generate-music-posts.mjs",
-    "test": "node --experimental-test-module-mocks --test scripts/hronir/lib/__tests__/ranking.test.js"
+    "test": "node --import tsx/esm --experimental-test-module-mocks --test scripts/hronir/lib/__tests__/ranking.test.js"
   },
   "dependencies": {
     "@astrojs/mdx": "^5.0.4",
